Additional Features
Going for a slightly different approach with Employee. Instead of
creating and then reacting to a invalidation event if anything goes
wrong, I am validating the Employee on receipt of the corresponding
EmployeeRegistered event. This is because there may be an issue with the
service sending the 'Invalidated' message, and then the consuming
process will be inconsistent with the main service. It also removed the
need for the 'Invalidated' messages. I may refactor bookings to do this
too.

If there was an issue in the above when creating booking, would this
retry/go to the error queue?

Don't like the property name 'ServiceValidated' but don't know what else
to call it just yet...

Sales has feature where you can add an employee. This is done by sending
the message and then
manually adding the employee to the list in JavaScript. I worry about
this because there could be an error creating the employee - it could
not be valid and then fail. Then Appointments could be added to an
employee that does not exist. Also, more business logic to register an
employee may
come along that may have to be replicated in Sales. Potential point of
business logic replication.

Have to change page to ajax on load consultant.

Probably a lot of refactoring to come.